added spammer detection method
each call of SpamDetection.checkUserRate(user) will keep the user
logged in a list.

if the user calls to many times the bot on a
limited time. (say 20 times in 10 minutes), the method will return
false.

it may not completely block an user, but prevent him to mak the bot
work too much.
for example if the user spams reactions in the Waitlist,
the list message won't update automatically anymore for him.
(but his status will still be updated in the db)